Mappers sin dependencia MApStruct, a la antigua

+++ src/main/java/com/alkemy/aceleracion/disney/dto/CharacterDTO.java
@@ -1,5 +1,6 @@
 package com.alkemy.aceleracion.disney.dto;
 
+import java.util.Collection;
 import java.util.Set;
 
 import lombok.Getter;
@@ -8,7 +9,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class CharacterDTO {
-	private int id;
+	private Long id;
     private String img;
     private String name;
     private int age;
